changes quantity in order to int

+++ webstore/src/main/java/com/packt/webstore/service/OrderService.java
@@ -2,5 +2,5 @@ package com.packt.webstore.service;
 
 public interface OrderService {
 	
-	void processOrder(String  productId, long quantity);
+	void processOrder(String  productId, int quantity);
 }
